DOPT - Determinant Optimality

The D-OPT problem for a given matrix $A \in \mathbb{R}^{m \times n}$ and a positive integer $s \le m$ is stated as:

\[\begin{array}{rl} \max & \log\det A' \cdot \text{diagm}(x) \cdot A \\ \text{s.t.} & x \in \mathbb{B}^{m} \\ & \sum_{i} x_{i} = s \end{array}\]

API

Solution Methods

DOPT.MetaHeuristic — Type

MetaHeuristic

The abstract class for defining new solution methods, to be used for triggering multiple dispatch over solve and init

Missing docstring.

Missing docstring for DOPT.AntColony. Check Documenter's build log for details.

DOPT.SimulatedAnnealing — Type

Simulated Annealing

Missing docstring.

Missing docstring for DOPT.LocalSearch. Check Documenter's build log for details.

DOPT.IteratedLocalSearch — Type

Iterated Local Search

DOPT.PathRelinking — Type

PathRelinking{M<:RelinkingMode} <: MetaHeuristic

DOPT.init — Function

init(A::Matrix{T})
init(A::Matrix{T}, R::Matrix{T})
init(::MetaHeuristic, A::Matrix{T})
init(::MetaHeuristic, A::Matrix{T}, R::Matrix{T})

DOPT.solve — Function

solve(::MetaHeuristic, A::Matrix{T}, s::Integer, args...; params...) where {T}
solve(::MetaHeuristic, A::Matrix{T}, R::Vector{Int}, s::Integer, args...; params...) where {T}

Solves the Determinant-Optimality problem for a given matrix $A \in \mathbb{R}^{m \times n}$.

Metrics

DOPT.gap — Function

gap(z::T, z̄::T)

Computes the relative gap between a solution $z$ and the reference value $\bar{z}$.

Rationale

Let

\[\begin{align*} z &= \log\det A' \cdot diagm(x) \cdot A \\ \bar{z} &= \log\det A' \cdot diagm(\bar{x}) \cdot A \end{align*}\]

Then, is is possible to compute the relative gap as

\[\begin{align*} z - \bar{z} &= \log\det A' \cdot diagm(x) \cdot A - \log\det A' \cdot diagm(\bar{x}) \cdot A \\ &= \log\frac{\det A' \cdot diagm(x) \cdot A}{\det A' \cdot diagm(\bar{x}) \cdot A} \\[2ex] 1 - e^{z - \bar{z}} &= \frac{\det A' \cdot diagm(\bar{x}) \cdot A - \det A' \cdot diagm(x) \cdot A}{\det A' \cdot diagm(\bar{x}) \cdot A} \end{align*}\]

Info

It is assumed that $\det A' \cdot diagm(\bar{x}) \cdot A \ge 0$. This computation comes from calling logabsdet.

Instances & Results

DOPT.read_instance — Function

read_instance(n::Integer, i::Integer)
read_instance(n::Integer, i::Integer, matrix_name::Symbol)

Example

julia> A, R = read_instance(200, 2);

or

julia> A = read_instance(200, 2, :A);

julia> R = read_instance(200, 2, :R);

DOPT.read_solution — Function

read_solution(n::Integer, i::Integer)

Missing docstring.

Missing docstring for DOPT.update_solution!. Check Documenter's build log for details.